@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
@@ -9,13 +9,13 @@
     <Protobuf Include="Protos\notification.proto" GrpcServices="Client"/>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.9.1" />
-    <PackageReference Include="Fluid.Core" Version="2.9.0" />
-    <PackageReference Include="Grpc.AspNetCore" Version="2.62.0" />
-    <PackageReference Include="Grpc.Tools" Version="2.63.0" PrivateAssets="all" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.4" />
-    <PackageReference Include="MinimalApis.LiquidViews" Version="2.9.0" />
-    <PackageReference Include="Serilog.AspNetCore" Version="8.0.1" />
+    <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.11.0"/>
+    <PackageReference Include="Fluid.Core" Version="2.16.0"/>
+    <PackageReference Include="Grpc.AspNetCore" Version="2.67.0"/>
+    <PackageReference Include="Grpc.Tools" Version="2.68.1" PrivateAssets="all"/>
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="9.0.0"/>
+    <PackageReference Include="MinimalApis.LiquidViews" Version="2.16.0"/>
+    <PackageReference Include="Serilog.AspNetCore" Version="9.0.0"/>
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\People.Infrastructure\People.Infrastructure.csproj"/>
