@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
@@ -8,13 +8,13 @@
     <ProjectReference Include="..\People.Domain\People.Domain.csproj"/>
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Confluent.Kafka" Version="2.4.0"/>
-    <PackageReference Include="MediatR" Version="12.2.0"/>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.4"/>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="8.0.4"/>
-    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="8.0.0"/>
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.2"/>
-    <PackageReference Include="Polly" Version="8.3.1"/>
-    <PackageReference Include="StackExchange.Redis" Version="2.7.33"/>
+    <PackageReference Include="Confluent.Kafka" Version="2.6.1"/>
+    <PackageReference Include="MediatR" Version="12.4.1"/>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.0"/>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="9.0.0"/>
+    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="9.0.0"/>
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.2"/>
+    <PackageReference Include="Polly" Version="8.5.0"/>
+    <PackageReference Include="StackExchange.Redis" Version="2.8.22"/>
   </ItemGroup>
 </Project>
