@@ -1,16 +1,16 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.4"/>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="8.0.4"/>
-    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="8.0.0"/>
-    <PackageReference Include="Microsoft.Extensions.Http.Resilience" Version="8.4.0"/>
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="8.0.2"/>
-    <PackageReference Include="Serilog.AspNetCore" Version="8.0.1"/>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.0"/>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="9.0.0"/>
+    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="9.0.0"/>
+    <PackageReference Include="Microsoft.Extensions.Http.Resilience" Version="9.0.0"/>
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.2"/>
+    <PackageReference Include="Serilog.AspNetCore" Version="9.0.0"/>
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\People.Kafka\People.Kafka.csproj"/>
