@@ -8,10 +8,8 @@ public sealed class GlobalExceptionHandler : IExceptionHandler
 {
     private readonly ILogger<GlobalExceptionHandler> _logger;
 
-    public GlobalExceptionHandler(ILogger<GlobalExceptionHandler> logger)
-    {
+    public GlobalExceptionHandler(ILogger<GlobalExceptionHandler> logger) =>
         _logger = logger;
-    }
 
     public async ValueTask<bool> TryHandleAsync(HttpContext context, Exception exception, CancellationToken ct)
     {
