@@ -5,5 +5,8 @@ namespace People.Grpc.People;
 public partial class EmailSigningInReply
 {
     internal static EmailSigningInReply Map(string token) =>
-        new() { Token = token };
+        new()
+        {
+            Token = token
+        };
 }
