@@ -6,5 +6,9 @@ namespace People.Grpc.People;
 public partial class SignUpReply
 {
     internal static SignUpReply Map(SignUpResult result) =>
-        new() { Id = result.Id, FullName = result.FullName };
+        new()
+        {
+            Id = result.Id,
+            FullName = result.FullName
+        };
 }
