@@ -4,14 +4,14 @@ using People.Domain.SeedWork;
 // ReSharper disable AutoPropertyCanBeMadeGetOnly.Local
 namespace People.Domain.Entities;
 
-public sealed class EmailAccount : Entity<Ulid>
+public sealed class EmailAccount : Entity<Guid>
 {
     private DateTime? _confirmedAt;
     private DateTime _createdAt;
 
     public EmailAccount(AccountId accountId, string email, bool isPrimary, DateTime? confirmedAt, DateTime createdAt)
     {
-        Id = Ulid.NewUlid();
+        Id = Guid.CreateVersion7();
         AccountId = accountId;
         Email = email;
         IsPrimary = isPrimary;
