@@ -1,7 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using People.Domain.Entities;
-using People.Infrastructure.ValueConverters;
 
 namespace People.Infrastructure.EntityConfigurations;
 
@@ -12,13 +11,16 @@ internal sealed class ExternalConnectionEntityTypeConfiguration : IEntityTypeCon
         builder.ToTable("connections");
 
         builder.HasKey(x => x.Id);
-        builder.HasAlternateKey(x => new { Name = x.Type, x.Identity });
+        builder.HasAlternateKey(x => new
+        {
+            Name = x.Type,
+            x.Identity
+        });
 
         builder.Ignore(x => x.DomainEvents);
 
         builder.Property(x => x.Id)
-            .HasColumnName("id")
-            .HasConversion<UlidConverter>();
+            .HasColumnName("id");
 
         builder.Property<AccountId>("_accountId")
             .HasColumnName("account_id")
